requirements.txt
openai
requests
pygame
pyyaml
python-dotenv
gradio
Pillow
PyQt6
